creating the joystick.

+++ app/src/main/java/com/example/flightgearandroidapp/viewModel/MyViewModel.java
@@ -2,6 +2,8 @@ package com.example.flightgearandroidapp.viewModel;
 
 import com.example.flightgearandroidapp.model.FGPlayer;
 import com.example.flightgearandroidapp.model.IFGPlayer;
+import com.example.flightgearandroidapp.views.Joystick;
+import com.example.flightgearandroidapp.views.JoystickListener;
 
 public class MyViewModel {
     private IFGPlayer modelFG;
@@ -19,7 +21,7 @@ public class MyViewModel {
     public void VNChangeThrottle(double throttle){
         modelFG.changeThrottle(throttle);
     }
-    public void VNChangeJoystick(double aileron, double elevator){
+    public void VNChangeJoystick(float aileron, float elevator){
         modelFG.changeJoystick(aileron,elevator);
     }
     public void VNChangeRudder(double rudder){
